Refactor some stuff & Fix ai flashcards loading indicator
