+++ apps/doc/src/annotations/annotation_popup/Actions/CreateAIFlashcard.tsx
@@ -6,8 +6,8 @@ import {useAnnotationPopupAction} from "../AnnotationPopupActionContext";
 import {IAnnotationPopupActionProps} from "../AnnotationPopupActions";
 
 export const CreateAIFlashcard: React.FC<IAnnotationPopupActionProps> = ({ annotation }) => {
-    const {clear} = useAnnotationPopupAction();
-    const [, handler] = useAutoFlashcardHandler(annotation);
+    const {clear, setAiFlashcardStatus} = useAnnotationPopupAction();
+    const [_, handler] = useAutoFlashcardHandler(annotation);
     const dialogs = useDialogManager();
 
     const verifiedAction = useAIFlashcardVerifiedAction();
@@ -15,9 +15,15 @@ export const CreateAIFlashcard: React.FC<IAnnotationPopupActionProps> = ({ annot
     React.useEffect(() => {
         clear();
         verifiedAction(() => {
+            setAiFlashcardStatus("waiting");
             handler()
-                .then(() => dialogs.snackbar({ message: "AI Flashcard created successfully!" }))
-                .catch(err => console.error("Could not handle verified action: ", err));
+                .then(() => {
+                    dialogs.snackbar({ message: "AI Flashcard created successfully!" });
+                    setAiFlashcardStatus("idle");
+                }).catch((e) => {
+                    console.error("Could not handle verified action: ", e);
+                    setAiFlashcardStatus("idle");
+                });
         });
     // eslint-disable-next-line react-hooks/exhaustive-deps
     }, []);
